use db to store the template library and clean up.
