+++ ckcore/core/query/template_expander.py
@@ -1,14 +1,12 @@
 import functools
 import json
 from abc import abstractmethod
-from typing import Any, Generator, ByteString, Union, Iterable
-from typing import Dict, Tuple
-from typing import List
-from typing import Optional
+from typing import Any, Generator, ByteString, Union, Iterable, Tuple, List, Optional
 
 from parsy import string, Parser, regex, any_char
 from ustache import default_getter, default_virtuals, render
 
+from core.db.templatedb import TemplateEntityDb
 from core.error import NoSuchTemplateError
 from core.parse_util import (
     double_quote_dp,
@@ -25,10 +23,73 @@ from core.parse_util import (
     equals_p,
     json_value_p,
 )
-from core.query import Expandable, TemplateExpander, Template, query_parser
-from core.query.model import Query
+from core.query import query_parser, QueryParser
+from core.query.model import Query, Expandable, Template
 from core.types import Json
 from core.util import identity
+
+
+class TemplateExpander(QueryParser):
+    """
+    TemplateExpander is able to maintain (crud) a set of templates
+    as well as expanding strings which might contain expandable sections that refer to templates
+    in the templates library.
+    """
+
+    @abstractmethod
+    async def expand(self, maybe_expandable: str) -> Tuple[str, List[Expandable]]:
+        """
+        Expand the given string, which might contain expandable sections.
+        All expandable sections get expanded and replaced in the string.
+        If there are no expandable sections, the string is returned as is.
+
+        If there are expandable sections, where the related template does not exist,
+        an exception is thrown.
+
+        :param maybe_expandable: a string which might contain expandable sections.
+        :return: A string with all expandable sections expanded.
+        :raises: NoSuchTemplateError if the related template of an expandable section does not exist.
+        """
+
+    @abstractmethod
+    def render(self, template: str, properties: Json) -> str:
+        """
+        Render a given template with given properties.
+        :param template: the template to render.
+        :param properties: the properties used to define the template properties.
+        :return: the fully rendered string without any template parameters.
+        """
+
+    @abstractmethod
+    async def put_template(self, template: Template) -> None:
+        """
+        Put a named template to the template library.
+        :param template: the template to put.
+        :return: None.
+        """
+
+    @abstractmethod
+    async def delete_template(self, name: str) -> None:
+        """
+        Delete a with given name from the template library.
+        :param name: the name of the template to delete.
+        :return: None.
+        """
+
+    @abstractmethod
+    async def get_template(self, name: str) -> Optional[Template]:
+        """
+        Return the template with the given name.
+        :param name: the name of the template.
+        :return: the template with the given name if the template exists otherwise None.
+        """
+
+    @abstractmethod
+    async def list_templates(self) -> List[Template]:
+        """
+        List all available templates in the system.
+        :return: all templates in the system.
+        """
 
 
 class TemplateExpanderBase(TemplateExpander):
@@ -54,9 +115,9 @@ class TemplateExpanderBase(TemplateExpander):
             return maybe_expandable, expands
 
     async def expand_part(self, expand: Expandable) -> str:
-        tpl = await self.template(expand.template)
+        tpl = await self.get_template(expand.template)
         if tpl:
-            return self.render(tpl, expand.props)
+            return self.render(tpl.template, expand.props)
         else:
             raise NoSuchTemplateError(expand.template)
 
@@ -68,33 +129,29 @@ class TemplateExpanderBase(TemplateExpander):
     def default_props(self) -> Optional[Json]:
         pass
 
-    @abstractmethod
-    async def template(self, name: str) -> Optional[str]:
-        pass
 
+class DBTemplateExpander(TemplateExpanderBase):
+    """
+    Template expander, which maintains the templates in the database.
+    """
 
-class InMemoryTemplateExpander(TemplateExpanderBase):
-    def __init__(self) -> None:
-        self.templates: Dict[str, Template] = {}
-        self.props: Json = {}
-
-    async def template(self, name: str) -> Optional[str]:
-        return self.templates[name].template if name in self.templates else None
-
-    async def put_template(self, template: Template) -> None:
-        self.templates[template.name] = template
-
-    async def delete_template(self, name: str) -> None:
-        self.templates.pop(name, None)
-
-    async def get_template(self, name: str) -> Optional[Template]:
-        return self.templates.get(name)
-
-    async def list_templates(self) -> List[Template]:
-        return list(self.templates.values())
+    def __init__(self, db: TemplateEntityDb) -> None:
+        self.db = db
 
     def default_props(self) -> Optional[Json]:
-        return self.props
+        return None
+
+    async def put_template(self, template: Template) -> None:
+        await self.db.update(template)
+
+    async def delete_template(self, name: str) -> None:
+        await self.db.delete(name)
+
+    async def get_template(self, name: str) -> Optional[Template]:
+        return await self.db.get(name)
+
+    async def list_templates(self) -> List[Template]:
+        return [t async for t in self.db.all()]
 
 
 def render_template(template: str, props: Json, more_props: Iterable[Json] = ()) -> str:
